fix: Fix behavior of card and section titles
